allow conftest functions to be imported as shiny.test module
