+++ shiny/test/__init__.py
@@ -1,0 +1,24 @@
+try:
+    import pytest_playwright  # noqa: F401 # pyright: ignore[reportUnusedImport, reportMissingTypeStubs]
+except ImportError:
+    raise ImportError(
+        "The shiny.test module requires the pytest-playwright package to be installed."
+        " Please install it with this command:"
+        "\n\n    pip install pytest-playwright"
+    )
+from playwright.sync_api import Page, expect
+
+
+from ._conftest import ShinyAppProc, create_app_fixture, local_app, run_shiny_app
+from ._expect import expect_to_change, retry_with_timeout
+
+__all__ = (
+    "Page",
+    "expect",
+    "ShinyAppProc",
+    "create_app_fixture",
+    "local_app",
+    "run_shiny_app",
+    "expect_to_change",
+    "retry_with_timeout",
+)
